refactor: 예약 상태 변경 기능과 sms 전송 기능 트랜잭션 분리
- 예약 변경과 sms 전송 로직이 하나의 트랜잭션에 묶여있어, sms 전송이 실패하면 상태 변경까지 롤백되는 상황이 문제였음. 두가지 로직을 파사드 객체에 위임하여 트랜잭션 분리
